fix: change datatypes

+++ src/datatypes/Array.java
@@ -4,24 +4,24 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Array {
-    private List<Integer> data;
+    private List<Double> data;
 
     public Array(int size) {
         data = new ArrayList<>(size);
         for (int i = 0; i < size; i++) {
-            data.add(0); 
+            data.add(0.0); 
         }
     }
 
-    public Array(List<Integer> data) {
+    public Array(List<Double> data) {
         this.data = data;
     }
 
-    public List<Integer> getData() {
+    public List<Double> getData() {
         return data;
     }
 
-    public void setData(List<Integer> data) {
+    public void setData(List<Double> data) {
         this.data = data;
     }
 
@@ -29,11 +29,11 @@ public class Array {
         return data.size();
     }
 
-    public Integer get(int index) {
+    public Double get(int index) {
         return data.get(index);
     }
 
-    public void set(int index, Integer value) {
+    public void set(int index, Double value) {
         data.set(index, value);
     }
 }
